prisoners
A ‘jeep’ (an org with a designated hitch) on eating an org will
dismember it and take its nodes prisoner.  A non-jeep will randomly
relocate the org if both are registered.

A prisoner is a node that is not a member of its org! Second class
citizen.

Burden now either shifted within an org, or a given node can
offloadBurden to the org members. Prisoners offload their burden to the
regular members of their org, to make them easier to tow. They also
donate all their photosynthesis, all the time, to the org head.